+++ src/main/java/edu/umn/ecology/populus/model/lpg/LPGParamInfo.java
@@ -274,7 +274,7 @@ public class LPGParamInfo implements BasicPlot {
             lagVector.addElement(y);
 
             if (kcounter >= lag) {
-                nx = lagVector.elementAt(0).doubleValue();
+                nx = lagVector.elementAt(0);
                 lagVector.removeElementAt(0);
             } else {
                 kcounter += dt;
